% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbs.R
\name{cbs}
\alias{cbs}
\title{Download 'live' cbs data}
\usage{
cbs(
  url = "https://opendata.cbs.nl/statline/#/CBS/nl/dataset/83913NED/table?dl=323FD"
)
}
\arguments{
\item{url}{url to cbs data}
}
\value{
data.frame
}
\description{
Enables the use of 'live' CBS data in your nicerplot
}
\details{
See manual for details.
}
\examples{
\dontrun{
link <- "https://opendata.cbs.nl/statline/#/CBS/nl/dataset/83913NED/table?dl=323FD"
head(cbs(link))

nplot(link)

j <- james(cbs(link))
j$title   <- "Stad nog steeds in trek"
j$y_title <- "verkoopprijzen (2015 = 100)"
nplot(j)
}
}
