% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/james-class.R
\name{is.james}
\alias{is.james}
\title{tests if its argument has class 'james'.}
\usage{
is.james(p)
}
\arguments{
\item{p}{an R object}
}
\value{
boolean
}
\description{
creates nice figures (png, pdf, svg, jpeg); to do: create example
}
