% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/james-class.R
\name{james}
\alias{james}
\title{Object to Save Data and Meta Data on Your Figure}
\usage{
james(...)
}
\arguments{
\item{...}{you can add parameters to customize your figure (see manual)}
}
\value{
object of class james
}
\description{
The main data object in the 'nicerplot' package has class `james`. `james` objects enable you to add data and meta data in an incremental fashion so you don't need to put everything in one and the same plot-statement.
}
