% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhelp.R
\name{nhelp}
\alias{nhelp}
\title{Shows help}
\usage{
nhelp(param = NULL)
}
\arguments{
\item{param}{the parameter for which you need help}
}
\description{
Opens manual. If param is given, it opens manual at description of given parameter.
}
\examples{
\dontrun{
nhelp("x_date_format")
}
}
