% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{nplot.data.frame}
\alias{nplot.data.frame}
\title{Nicer Plot of Data Frame}
\usage{
nplot.data.frame(x, ...)
}
\arguments{
\item{x}{data.frame}

\item{...}{you can add parameters to customize your figure (see manual)}
}
\value{
path/to/result/files.png
}
\description{
Creates nice figures (PNG, PDF, SVG, JPEG).
}
