% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{nplot.james}
\alias{nplot.james}
\title{Nicer Plot of James Object}
\usage{
nplot.james(p, ...)
}
\arguments{
\item{p}{object of class james}

\item{...}{you can add parameters to customize your figure (see manual)}
}
\value{
path/to/result/files.png
}
\description{
Creates nice figures (PNG, PDF, SVG, JPEG).
}
