% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{nplot.matrix}
\alias{nplot.matrix}
\title{Nicer Plot of Matrix}
\usage{
nplot.matrix(x, ...)
}
\arguments{
\item{x}{matrix}

\item{...}{you can add parameters to customize your figure (see manual)}
}
\value{
path/to/result/files.png
}
\description{
Creates nice figures (PNG, PDF, SVG, JPEG).
}
