% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nplot.R
\name{nplot}
\alias{nplot}
\title{Generic function to make a 'nice R plot'/'nicer plot` and save it as png, pdf, jpg and/or svg.}
\usage{
nplot(x, ...)
}
\arguments{
\item{x}{'path/to/file.xlsx', 'URL', or above data types.}

\item{...}{you can add parameters to customize your figure (see manual). For example `title` to specify a title, `pdf = TRUE` if you want a PDF (`png = TRUE` by default), and `file` to specify file name (with `file`.{png,pdf,jpg,svg} as a result).}
}
\value{
path/to/result/file.png
}
\description{
Can plot time series data from xlsx-files, certain URLs, and directly from R (`data.frame`, `james`, `list`, `matrix`, `mts`, `ts`). Saves it as png, pdf, svg, and/or jpeg.
}
\examples{
\dontrun{
x <- 0:6
my_data <- data.frame(x, first = (6 - x)^2, second = x^2)
nplot(my_data, title = 'Hello World', x_title = 'x', y_title = 'y', footnote = "just an example")
}
}
