% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{nplot.ts}
\alias{nplot.ts}
\title{Nicer Plot of Time Series Data}
\usage{
nplot.ts(x, ...)
}
\arguments{
\item{x}{ts object}

\item{...}{you can add parameters to customize your figure (see manual)}
}
\value{
path/to/result/files.png
}
\description{
Creates nice figures (PNG, PDF, SVG, JPEG).
}
