% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.data.frame}
\alias{plot.data.frame}
\title{Plots data in data frame}
\usage{
plot.data.frame(x, ...)
}
\arguments{
\item{x}{data.frame}

\item{...}{you can add parameters to customize your figure (see manual)}
}
\value{
path/to/result/files.png
}
\description{
Creates nice figures (png, pdf, svg, jpeg).
}
