% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.james}
\alias{plot.james}
\title{plots data in so-called 'james' object}
\usage{
\method{plot}{james}(p, ...)
}
\arguments{
\item{p}{object of class james}

\item{...}{you can add parameters to customize your figure (see manual)}
}
\value{
path/to/result/files.png
}
\description{
creates nice figures (png, pdf, svg, jpeg); to do: create example
}
