% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.matrix}
\alias{plot.matrix}
\title{plots data in matrix}
\usage{
\method{plot}{matrix}(mat, ...)
}
\arguments{
\item{mat}{matrix}
}
\value{
path/to/result/files.png
}
\description{
creates nice figures (png, pdf, svg, jpeg); to do: create example
}
