% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.ts}
\alias{plot.ts}
\title{plots data in ts object}
\usage{
\method{plot}{ts}(x, ...)
}
\arguments{
\item{x}{ts object}

\item{...}{you can add parameters to customize your figure (see manual)}
}
\value{
path/to/result/files.png
}
\description{
creates nice figures (png, pdf, svg, jpeg); to do: create example
}
