% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/james-class.R
\name{print.james}
\alias{print.james}
\title{Print James Data Object in Nice Fashion}
\usage{
\method{print}{james}(x, ...)
}
\arguments{
\item{x}{object of class `james`}

\item{...}{further arguments are ignored. The ... argument was added to be consistent with the generic print function}
}
\description{
Print James Data Object in Nice Fashion
}
\details{
Relevant parameters are shown with values that are properly aligned. Data are truncated.
}
\examples{
x <- 0:6
my_data <- data.frame(x, first = (6 - x)^2, second = x^2)
p <- james(data = my_data, title = "Hello World")
p
# `print(p)` and `print.james(p)` give same results
}
