% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/james-class.R
\name{is.james}
\alias{is.james}
\title{Tests if its Argument has Class 'james'.}
\usage{
\method{is}{james}(p)
}
\arguments{
\item{p}{an R object}
}
\value{
boolean
}
\description{
`james` objects enable you to add data and meta data in an incremental fashion so you don't need to put everything in one and the same plot-statement. With `is.james` you can test whether a given object is of type `james`.
}
