% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is_really_character}
\alias{is_really_character}
\title{Indicates whether all values in vec are really char}
\usage{
is_really_character(vec)
}
\arguments{
\item{vec}{vector}
}
\description{
Indicates whether all values in vec are really char
}
\keyword{internal}
