% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/james-class.R
\name{james}
\alias{james}
\title{nicerplot object}
\usage{
james(...)
}
\arguments{
\item{...}{may be list or individual values, i.e. james(list(a=1,b=2)) or james(a=1,b=2) or james()}
}
\value{
object of class james
}
\description{
creates object you can plot
}
