% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/james-class.R
\name{james}
\alias{james}
\title{Object to Save Data and Meta Data on Your Figure}
\usage{
james(...)
}
\arguments{
\item{...}{add parameters to customize figure (see manual)}
}
\value{
object of class james
}
\description{
The main data object in the 'nicerplot' package has class `james`. `james` objects enable you to add data and meta data in an incremental fashion so you don't need to put everything in one and the same plot-statement.
}
\details{
Knowing the right parameter names is crucial. One example is parameter 'data', which
you can use to store your data object. Please see manual for details. Please find the
manual here: \url{https://data-science-made-easy.github.io/nicerplot}.
}
\examples{
\dontrun{
x <- 0:6
d <- data.frame(x, first = (6 - x)^2, second = x^2)
j <- nicerplot::james(data = d, title = 'Hello World', x_title = 'x', y_title = 'y') # init
j$footnote <- 'just an example' # add other parameters
nicerplot::nplot(j)
}
}
