% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/james-class.R
\name{james}
\alias{james}
\title{nicerplot object}
\usage{
james(...)
}
\arguments{
\item{...}{you can add parameters to customize your figure (see manual)}
}
\value{
object of class james
}
\description{
creates object you can plot
}
