% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nplot.R
\name{nplot}
\alias{nplot}
\title{Create a 'nice R plot'/'nicer plot` in Png, Pdf, Jpg and/or Svg format}
\usage{
nplot(x, ...)
}
\arguments{
\item{x}{data.frame, james object, matrix, mts, ts, 'path/to/file.xlsx', URL}

\item{...}{you can add parameters to customize your figure (see manual). For example `title` to specify a title, `pdf = TRUE` if you want a PDF (`png = TRUE` by default), and `file` to specify file name (with path/to/your-file.\{png, pdf, jpg, svg\} as a result).}
}
\value{
path/to/result/file.png
}
\description{
Plots data stored in a data.frame, james object, matrix, mts, ts, cbs.nl-type 'share data' URL,
an xlsx-file or in a list that combines these data structures. The result is saved as
PNG, PDF, SVG, and/or JPEG.
}
\details{
Class 'james' can help you to recycle parameters and data accross multiple figures. A 'james'
object also enables you also to add parameters in an incremental fashion, rather than
putting them all in one big nplot-statement.
}
\examples{
\dontrun{
x <- 0:6
d <- data.frame(x, first = (6 - x)^2, second = x^2)
nplot(d, title = 'Hello World', x_title = 'x', y_title = 'y', footnote = "just an example")
}
}
\seealso{
\code{\link{james}} to create a james object
}
