% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.character}
\alias{plot.character}
\title{plots data from file or URL}
\usage{
\method{plot}{character}(string, ...)
}
\arguments{
\item{string}{may be 'path/to/file.xlsx' (please note: at the current moment Excel files must have a meta-tab) or a URL refering to cbs.nl}
}
\value{
path/to/result/files.png
}
\description{
creates nice figures (png, pdf, svg, jpeg); to do: create example
}
