% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.data.frame}
\alias{plot.data.frame}
\title{plots data in data frame}
\usage{
\method{plot}{data.frame}(mat, ...)
}
\arguments{
\item{mat}{data.frame}
}
\value{
path/to/result/files.png
}
\description{
creates nice figures (png, pdf, svg, jpeg); to do: create example
}
