% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.list}
\alias{plot.list}
\title{Plots data in list}
\usage{
plot.list(x, ...)
}
\arguments{
\item{x}{list must contain elements of a class that can be plotted by this package (e.g., matrix, data.frame, character, ts, mts)}

\item{...}{you can add parameters to customize your figure (see manual)}
}
\value{
path/to/result/files.png
}
\description{
Creates nice figures (PNG, PDF, SVG, JPEG).
}
