% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.matrix}
\alias{plot.matrix}
\title{Plots data in matrix}
\usage{
plot.matrix(x, ...)
}
\arguments{
\item{x}{matrix}

\item{...}{you can add parameters to customize your figure (see manual)}
}
\value{
path/to/result/files.png
}
\description{
Creates nice figures (PNG, PDF, SVG, JPEG).
}
