% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.mts}
\alias{plot.mts}
\title{Plots data in mts object}
\usage{
plot.mts(x, ...)
}
\arguments{
\item{x}{ts object}

\item{...}{you can add parameters to customize your figure (see manual)}
}
\value{
path/to/result/files.png
}
\description{
Creates nice figures (png, pdf, svg, jpeg).
}
