% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.ts}
\alias{plot.ts}
\title{Plots data in ts object}
\usage{
plot.ts(x, ...)
}
\arguments{
\item{x}{ts object}

\item{...}{you can add parameters to customize your figure (see manual)}
}
\value{
path/to/result/files.png
}
\description{
Creates nice figures (PNG, PDF, SVG, JPEG).
}
